changed seasoning record type for menu struct optimization

+++ src/main/kotlin/com/menuplanner/server/menu/entity/SeasoningRecord.kt
@@ -10,7 +10,7 @@ data class SeasoningRecord(
     override val seasoningId: Int = 0,
     override val id: Int = 0,
     override val item: String = "",
-    override val quantity: Int = 0,
+    override val quantity: Double = 0.0,
     override val scale: String = "",
 ) : Seasoning by SeasoningImpl()
 
@@ -18,7 +18,7 @@ class SeasoningImpl : Seasoning {
     override val seasoningId: Int = 0
     override val id: Int = 0
     override val item: String = ""
-    override val quantity: Int = 0
+    override val quantity: Double = 0.0
     override val scale: String = ""
 }
 
@@ -26,6 +26,6 @@ interface Seasoning {
     val seasoningId: Int
     val id: Int
     val item: String
-    val quantity: Int
+    val quantity: Double
     val scale: String
 }
